extract wallet connection to new component
